parser/vk/requirements.txt
pandas==3.0.0
python-dotenv==1.2.1
vk_api==11.10.0
